% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la.anova.R
\name{la.anova}
\alias{la.anova}
\alias{laWRBM.anova}
\alias{laBRBM.anova}
\title{MRMC Analysis of Limits of Agreement using ANOVA}
\usage{
laWRBM.anova(
  df,
  modalitiesToCompare = c("testA", "testB"),
  keyColumns = c("readerID", "caseID", "modalityID", "score"),
  if.aov = TRUE
)

laBRBM.anova(
  df,
  modalitiesToCompare = c("testA", "testB"),
  keyColumns = c("readerID", "caseID", "modalityID", "score"),
  if.aov = TRUE
)
}
\arguments{
\item{df}{Data frame of observations, one per row. Columns identify random effects, fixed effects,
and the observation. Namely,
\describe{
  \item{readerID}{The factor corresponding to the different readers in the study.
    The readerID is treated as a random effect.}
  \item{caseID}{The factor corresponding to the different cases in the study.
    The caseID is treated as a random effect.}
  \item{modalityID}{The factor corresponding to the different modalities in the study.
    The modalityID is treated as a fixed effect.}
  \item{score}{The score given by the reader to the case for the modality indicated.}
}}

\item{modalitiesToCompare}{The factors identifying the modalities to compare. It should be at length 2. Default
\code{modalitiesToCompare = c("testA","testB")}}

\item{keyColumns}{Identify the factors corresponding to the readerID, caseID, modalityID, and score
(or alternative random and fixed effects). Default \code{keyColumns = c("readerID", "caseID",
"modalityID", "score")}}

\item{if.aov}{Boolean value to determine whether using aov function to do ANOVA. Default \code{if.aov = TRUE}}
}
\value{
A dataframe with one row. Each column is as following:
\describe{
  \item{meanDiff}{The mean of difference score.}
  \item{var.MeanDiff}{The variance of mean difference score}
  \item{var.1obs}{The variance of a single WRBM/BRBM difference score}
  \item{ci95meanDiff.bot}{Lower bound of 95\% CI for the mean difference score. \code{meanDiff+
  1.96*sqrt(var.MeanDiff)}}
  \item{ci95meanDiff.top}{Upper bound of 95\% CI for the mean difference score. \code{meanDiff-
  1.96*sqrt(var.MeanDiff)}}
  \item{la.bot}{Lower bound of WRBM/BRBM Limits of Agreement. \code{meanDiff+2*sqrt(var.1obs)}}
  \item{la.top}{Upper bound of WRBM/BRBM Limits of Agreement. \code{meanDiff-2*sqrt(var.1obs)}}
}

The two function shows the same 95\% CI for the mean difference score, but difference Limits of Agreements.
}
\description{
These two functions calculate two types of Limits of Agreement using ANOVA: Within-Reader Between-Modality(WRBM)
and Between-Reader Between-Modality(BRBM). The 95\% confidence interval of the mean difference is also provided.
The ANOVA method are realized either by applying stats::aov or by matrix multiplication. See more details below
about the model structure.
}
\details{
Suppose the score from reader j for case k under modality i is\eqn{X_{ijk}}, then the difference score from the
same reader for the same cases under two different modalities is \eqn{Y_{jk} = X_{1jk} - X_{2jk}}.
\itemize{
  \item\code{laWRBM} use two-way random effect ANOVA to analyze the difference scores \eqn{Y_{jk}}. The model
  is \eqn{Y_{jk}=\mu + R_j + C_k + \epsilon_{jk}}, where \eqn{R_j} and \eqn{C_k} are random effects for readers
  and cases. The variance of mean and individual observation is expressed as the linear combination of the MS
  given by ANOVA.
  \item\code{laBRBM} use three-way mixed effect ANOVA to analyze the scores \eqn{X_{ijk}}. The model is given by
  \eqn{X_{ijk}=\mu + R_j + C_k + m_i + RC_{jk} + mR_{ij} + mC_{ik} + \epsilon_{ijk}}, where \eqn{R_j} and
  \eqn{C_k} are random effects for readers and cases and \eqn{m_i} is a fixed effect for modality. The variance
  of mean and individual observation is expressed as the linear combination of the MS given by ANOVA.
}
}
\examples{
config <- sim.NormalIG.Hierarchical.config(modalityID = c("testA", "testB"))
# Simulate an MRMC ROC data set
dFrame <- sim.NormalIG.Hierarchical(config)

# Compute Limits of Agreement
laWRBM_result <- laWRBM.anova(dFrame)
print(laWRBM_result)
laBRBM_result <- laBRBM.anova(dFrame)
print(laBRBM_result)

}
