% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{result.validSimulation.alphaR}
\alias{result.validSimulation.alphaR}
\alias{result.validSimulation.sigmaC}
\title{{
This is a data frame with the result from the simulation study to verify the simulation is consistent with the derived theoretical values
}}
\format{
A dataframe with the following column:
\describe{
  \item{sigma_C}{Case related parameter}
  \item{alpha_R}{Reader related parameter}
  \item{WR_mean}{Monte Carlo mean of the WRBM difference score}
  \item{WR_var}{Monte Carlo variance of the WRBM difference score}
  \item{BR_mean}{Monte Carlo mean of the BRBM difference score}
  \item{BR_var}{Monte Carlo variance of the BRBM difference score}
  \item{True_WR_var}{Theoretical value for WRBM variance derived from the simulation model}
  \item{True_BR_var}{Theoretical value for BRBM variance derived from the simulation model}
  \item{WR_var_relative_bias}{relative bias for WRBM variance}
  \item{BR_var_relative_bias}{relative bias for BRBM variance}
}

An object of class \code{data.frame} with 20 rows and 10 columns.
}
\usage{
result.validSimulation.alphaR

result.validSimulation.sigmaC
}
\description{
{
This is a data frame with the result from the simulation study to verify the simulation is consistent with the derived theoretical values
}
}
\details{
There are two parameter sets for this simulation:
\describe{
  \item{result.validSimulation.alphaR}{
    In this study, we fixed the case-related parameter but change the reader related fixed the case related
    parameters \eqn{\sigma_C^2=\sigma_{\tau C}^2=1} and allowed the reader related parameter \eqn{\alpha_R
    (=\alpha_{\tau R})} to range from 2 to 20.
  }
  \item{result.validSimulation.sigmaC}{
    In this study, we fixed the reader related parameters \eqn{\alpha_R=\alpha_{\tau R}=10} and allowed the
    case related parameter \eqn{\sigma_C^2 (=\sigma_{\tau C}^2)} to range from 0.1 to 2, incrementing by 0.1.
  }
}
In both parameter sets, \eqn{\beta_R=\beta_{\tau R}=1}, so that the reader variability is only affected by
\eqn{\alpha_R}. We simulated 100,000 trials with each trial having 4 measurements from 2 readers for a single
case under 2 modalities. The Monte Carlo estimates of the variances are the sample variances of the 100,000
independent WRBM and BRBM differences.
}
\references{

}
\keyword{datasets}
