% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateMRMCVarEstimation.R
\name{validateMRMCVarEstimation}
\alias{validateMRMCVarEstimation}
\title{Validate and characterize the MRMC limits of agreement estimates}
\usage{
validateMRMCVarEstimation(
  nR.list,
  nC.list,
  alpha_R.list,
  sigma_C.list,
  nTrials = 1000
)
}
\arguments{
\item{nR.list}{A \eqn{r}-dimension array for the number of readers in each MRMC study}

\item{nC.list}{A \eqn{c}-dimension array for the number of cases in each MRMC study}

\item{alpha_R.list}{A \eqn{a}-dimension array for the reader related parameter}

\item{sigma_C.list}{A \eqn{b}-dimension array for the case related parameter}

\item{nTrials}{Number of MRMC simulations. Default is 1000}
}
\value{
A dataframe with \eqn{r*c*a*b} rows. Each column is as following:
\describe{
  \item{nReader}{Number of Readers}
  \item{nCase}{Number of Cases}
  \item{alpha_R}{Reader related parameter}
  \item{sigma_C}{Case related parameter}
  \item{WR_var_MCmean}{Monte Carlo mean of variance estimation for the WRBM difference score}
  \item{WR_var_MCvar}{Monte Carlo variance of variance estimation for the WRBM difference score}
  \item{True_WR_var}{Theoretical value for WRBM variance derived from the simulation model}
  \item{WR_var_relative_bias}{relative bias for WRBM variance}
  \item{WR_CV}{coefficient of variation for WRBM variance}
  \item{BR_var_MCmean}{Monte Carlo mean of variance estimation for the BRBM difference score}
  \item{BR_var_MCvar}{Monte Carlo variance of variance estimation for the BRBM difference score}
  \item{True_BR_var}{Theoretical value for BRBM variance derived from the simulation model}
  \item{BR_var_relative_bias}{relative bias for BRBM variance}
  \item{BR_CV}{coefficient of variation for BRBM variance}
  }
}
\description{
This function simulates MRMC study with \code{nReader} readers and \code{nCase} cases. From the output of \code{laWRBM.anova}
and \code{laBRBM.anova}, we exact the variance estimations for WRBM and BRBM differences. It outputs the Monte Carlo
mean of the variance estimation across all the simulation trials. The relative bias of the Monte Carlo mean comparing
to the theoretical value of the variance and the coefficient of variation is also calculated.
}
\details{
Let \eqn{\hat{V}_{BR}^{12}} denote the estimation of the variance for the BRBM difference for each simulated MRMC study, then
the relative bias is defined as \eqn{relativeBias~(\hat{V}_{BR}^{12})=(\sum{\hat{V}_{BR}^{12}}/{nTrials}-V_{BR}^{12})/V_{BR}^{12}},
and the coefficient of variantion is defined as \eqn{CV~(\hat{V}_{BR}^{12})=sd(\hat{V}_{BR}^{12})/{V_{BR}^{12}}}, where \eqn{sd()}
denote sample standard deviation.
}
\examples{
library(iMRMC)
nR.list <- c(5)
nC.list <- c(100)
sigma_C.list <- c(1)
alpha_R.list <- c(10)
#result <- validateMRMCVarEstimation(nR.list, nC.list, alpha_R.list, sigma_C.list)

}
